+++ docs/changelog.rst
@@ -12,8 +12,6 @@ Unreleased
 New features
 ~~~~~~~~~~~~
 
-- :doc:`plugins/lastgenre`: Added ``cleanup_existing`` configuration flag to
-  allow whitelist canonicalization of existing genres.
 - Add native support for multiple genres per album/track. The ``genres`` field
   now stores genres as a list and is written to files as multiple individual
   genre tags (e.g., separate GENRE tags for FLAC/MP3). The
@@ -45,6 +43,8 @@ New features
   in ``lyrics_language`` and translation language (if translations present) in
   ``lyrics_translation_language`` flexible attributes. Lyrics are automatically
   migrated on the first beets run. :bug:`6370`
+- :doc:`plugins/lastgenre`: Added ``cleanup_existing`` configuration flag to
+  allow whitelist canonicalization of existing genres.
 
 Bug fixes
 ~~~~~~~~~
@@ -76,6 +76,9 @@ For plugin developers
   please update it to populate a list of ``genres`` instead. You will see a
   deprecation warning for now, but support for populating the single ``genre``
   field will be removed in version ``3.0.0``.
+- The ``tunelog`` logging helper that was exclusively available to the lastgenre
+  plugin is now usable througout beets and was renamed to ``extra_debug``.
+  Import it from the ``beets.logging`` module to use it.
 
 Other changes
 ~~~~~~~~~~~~~